[DM-16383] Get everything working at NCSA.
Add a check to determine this and store the hostname in an
environment variable if we're at NCSA.  If so, don't create
the mock-qserv (use the real one) and use the right ingress
rule depending on where we are.

+++ kube/destroy.sh
@@ -1,4 +1,8 @@
 #!/bin/bash -x
+# If the hostname we're deploying from is at NCSA,
+# this string will not be empty, allowing for us to
+# do different things for GKE / NCSA.
+NCSA_DEPLOY=`hostname -f | grep ncsa`
 
 # Delete the Oracle backend.
 kubectl delete deployment/oracle-deployment
@@ -8,9 +12,11 @@ kubectl delete service/oracle-service
 kubectl delete deployment/tap-schema-deployment
 kubectl delete service/tap-schema-service
 
-# Delete the mock qserv backend.
-kubectl delete deployment/mock-qserv-deployment
-kubectl delete service/qserv-master01
+if [ -z "$NCSA_DEPLOY" ]; then
+  # Delete the mock qserv backend.
+  kubectl delete deployment/mock-qserv-deployment
+  kubectl delete service/qserv-master01
+fi
 
 # Delete the postgres backend for UWS jobs.
 kubectl delete service/postgresql-service
